requirements.txt
ansible==2.7.5
asn1crypto==0.24.0
asyncssh==1.16.1
backcall==0.1.0
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
cryptography==2.6.1
decorator==4.4.0
dill==0.2.9
docopt==0.6.2
Flask==1.0.2
future==0.17.1
genie==19.0.1
genie.abstract==19.0.0
genie.conf==19.0.1
genie.examples==19.0.1
genie.harness==19.0.2
genie.libs.conf==19.0
genie.libs.filetransferutils==19.0.0
genie.libs.ops==19.0.2
genie.libs.parser==19.0.3
genie.libs.sdk==19.0
genie.libs.telemetry==19.0.0
genie.metaparser==19.0
genie.ops==19.0.1
genie.parsergen==19.0
genie.predcore==19.0.0
genie.telemetry==19.0.2
genie.trafficgen==19.0
genie.utils==19.0.1
gitdb2==2.0.5
GitPython==2.1.11
gunicorn==19.9.0
idna==2.8
ipyats==0.2.0
ipython==7.4.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonpickle==1.1
junit-xml==1.8
junos-eznc==2.2.0
lxml==4.3.0
MarkupSafe==1.1.1
napalm==2.3.3
napalm-ansible==0.10.0
napalm-asa==0.1.1
napalm-vyos==0.1.6
ncclient==0.6.3
netaddr==0.7.19
netmiko==2.3.0
paramiko==2.4.2
parso==0.4.0
pathspec==0.5.9
pexpect==4.7.0
pickleshare==0.7.5
prettytable==0.7.2
prompt-toolkit==2.0.9
psutil==5.6.1
ptyprocess==0.6.0
pyasn1==0.4.5
pyats==19.0
pyats.aereport==19.0
pyats.aetest==19.0.1
pyats.async==19.0
pyats.connections==19.0
pyats.datastructures==19.0
pyats.easypy==19.0.1
pyats.examples==19.0
pyats.kleenex==19.0.1
pyats.log==19.0.1
pyats.results==19.0
pyats.robot==19.0.1
pyats.tcl==19.0
pyats.templates==19.0
pyats.topology==19.0
pyats.utils==19.0.1
pycparser==2.19
pyeapi==0.8.2
Pygments==2.3.1
pyIOSXR==0.53
PyNaCl==1.3.0
pynetbox==4.0.6
pynxos==0.0.3
pyserial==3.4
python-dateutil==2.7.5
PyYAML==3.13
requests==2.21.0
robotframework==3.1.1
robotframework-sshlibrary==3.3.0
scp==0.13.0
selectors2==2.0.1
setproctitle==1.1.10
six==1.12.0
smmap2==2.0.5
tabulate==0.8.2
textfsm==0.4.1
tqdm==4.31.1
traitlets==4.3.2
unicon==19.0.2
urllib3==1.24.1
virlutils==0.8.4
VyattaConfParser==0.5.1
wcwidth==0.1.7
Werkzeug==0.14.1
xmltodict==0.12.0
yamllint==1.15.0
